step 4 of catmull clark works

+++ godot-3.2.2-stable/godot-3.2.2-stable/modules/catmullclark/trimesh.cpp
@@ -271,8 +271,8 @@ std::vector<vertex> trimesh_t::getAllVertexPositions()
 
 vertex getAverage(std::vector<vertex> verteces)
 {
-	int totalX = 0, totalY = 0;
-	int numOfEntries = verteces.size();
+	float totalX = 0, totalY = 0;
+	float numOfEntries = verteces.size();
 
 	if (numOfEntries == 0)
 	{
